+++ src/generators/promptadapter.py
@@ -1,6 +1,5 @@
 from abc import ABC, abstractmethod
 
-
 class PromptAdapter(ABC):
     @abstractmethod
     def __init__(self, prompt: str = "") -> None:
