+++ src/generators/streamlitappgenerator.py
@@ -1,46 +1,9 @@
+from .app_generator import AppGenerator
 import time
 import tempfile
 import sys
 import subprocess
-import streamlit as st
-from abc import ABC, abstractmethod
-from .llm_client import SiemensLLMClient, WorkstationLLMClient
-from .prompt import StreamlitAppPromptAdapter, IEAppPromptAdapter
-
-
-# Base class for LLM clients
-class AppGenerator(ABC):
-    @abstractmethod
-    def __init__(self, llm_model="LLaMA-3-Latest", api_key=""):
-        self.prompt = ""
-        self.select_llm_client(llm_model, api_key)
-
-    def select_llm_client(self, llm_model, api_key=""):
-        """Select the appropriate LLM client based on user choice."""
-        if llm_model == "Siemens LLM":
-            if api_key:
-                api_key = api_key
-                self.llm_client = SiemensLLMClient(api_key)
-            else:
-                raise Exception("No API Key.")
-        else:
-            self.llm_client = WorkstationLLMClient(llm_model)
-
-    def get_requirements(self, prompt):
-        pass
-
-    def run_pipeline(self, prompt):
-        pass
-
-    def preview(self, code):
-        pass
-
-    def deploy(self, code):
-        pass
-
-    def stop(self):
-        pass
-
+from .promptadapter import StreamlitAppPromptAdapter
 
 class StreamlitAppGenerator(AppGenerator):
     def __init__(self, llm_model="LLaMA-3-Latest", api_key=""):
@@ -96,19 +59,3 @@ class StreamlitAppGenerator(AppGenerator):
             st.session_state.process.terminate()
             st.session_state.process = None
             st.success("App running stopped.")
-
-
-class IEAppGenerator(AppGenerator):
-    def __init__(self, llm_model="LLaMA-3-Latest", api_key=""):
-        super().__init__(llm_model, api_key)
-        self.prompt_adapter = IEAppPromptAdapter()
-        self.requirements = ""
-
-    def get_requirements(self):
-        self.requirements = self.llm_client.get_response(
-            self.prompt_adapter.requirement_prompt()
-        )
-
-    def run_pipeline(self, prompt):
-        self.prompt_adapter.update_user_prompt(prompt)
-        self.get_requirements()
