+++ src/prompts/define_task_distribution.prompt
@@ -1,0 +1,3 @@
+You are a software engineer working on the detailed design of a software architecture comprising two components. The backend program is responsible for communication with shop floor machines via MQTT. Furthermore, it is responsible for data processing and machine control. The frontend serves as the user interface and displays data, but does not process any data itself. It also may offer controls to the machines. The exact behavior of the frontend and the backend depends on the use-case description and software requirements provided below. Here is what I want you to do: Starting with the backend program: - summarize its task given the specific use-case. - define the data flow. Specify what data needs to be calculated/processed. Specify how to calculate this data. - summarize the interaction between the backend and the shop floor machine via MQTT. Include specific details including the data exchanged, its format, and the topic name. Continuing with the frontend program: - summarize its task including the interface layout. - specify what data needs to be queried from the backend. Finally, describe the interaction between the frontend and backend (what data needs to be exchanged, how often). Do not specify API endpoints, as these will be determined later. Also do not define any further details about the technologies than already provided. Do not provide any further recommendation of possible future features. Do not provide a rationale to your response. Make sure all information in the use-case and software requirements are included in your response, because we will only use your response in the future for development. Make sure no information is lost. Here is the use-case description and software requirements:
+
+[use case]
