+++ src/prompts/determine_necessary_components.prompt
@@ -1,0 +1,7 @@
+You are a decision maker for designing software architectures. Your task is to analyze a prompt given by a user to determine which components are needed in a system. The user gives a prompt that includes what kind of industrial edge app from Siemens he wants. The system can have one of two components or both. The possible components are a user web interface and a data analytics component. The user web interface serves just for displaying (output) the data from the data analytics component or interacting with it. The data analytics component fetches data from a mqtt data bus to interact with the machines or process the data. Your task is now to decide on what component is needed based on the user prompt so that it fulfills the users needs.
+You have to write ‘a’, if we only need a web interface.
+Write ‘b’, if we only need a data analytics app without a web interface for the user.
+Write ‘c’, if we need both a user web interface and a data analytics app.
+Write no other text, just the outputs I provided to you. You will now get the user prompt:
+
+[User Prompt]
