--- src/utils/streamlit_tools.py
@@ -1,10 +1,0 @@
-from langchain_core.tools import tool
-import streamlit as st
-
-
-@tool
-def createTimer() -> None:
-    """Creates a streamlit timer
-    """
-    pass
-
